src/models/requirements.txt
torch
torchvision
gradio
Pillow
scikit-learn
numpy
matplotlib
seaborn